Reorganize util
