+++ src/common/util/AndroidTimerFix.js
